Lagt til UnleashToggle for gjennomforing v2

+++ src/main/kotlin/no/nav/amt/deltaker/bff/deltakerliste/kafka/DeltakerlisteConsumer.kt
@@ -1,12 +1,14 @@
 package no.nav.amt.deltaker.bff.deltakerliste.kafka
 
 import com.fasterxml.jackson.module.kotlin.readValue
+import no.nav.amt.deltaker.bff.Environment
 import no.nav.amt.deltaker.bff.arrangor.ArrangorService
 import no.nav.amt.deltaker.bff.auth.TilgangskontrollService
 import no.nav.amt.deltaker.bff.deltaker.PameldingService
 import no.nav.amt.deltaker.bff.deltakerliste.Deltakerliste
 import no.nav.amt.deltaker.bff.deltakerliste.DeltakerlisteRepository
 import no.nav.amt.deltaker.bff.deltakerliste.tiltakstype.TiltakstypeRepository
+import no.nav.amt.deltaker.bff.unleash.UnleashToggle
 import no.nav.amt.deltaker.bff.utils.KafkaConsumerFactory.buildManagedKafkaConsumer
 import no.nav.amt.lib.kafka.Consumer
 import no.nav.amt.lib.models.deltakerliste.tiltakstype.Tiltakskode
@@ -20,6 +22,7 @@ class DeltakerlisteConsumer(
     private val tiltakstypeRepository: TiltakstypeRepository,
     private val pameldingService: PameldingService,
     private val tilgangskontrollService: TilgangskontrollService,
+    private val unleashToggle: UnleashToggle,
     private val topic: String,
 ) : Consumer<UUID, String?> {
     private val log = LoggerFactory.getLogger(javaClass)
@@ -34,6 +37,10 @@ class DeltakerlisteConsumer(
     override suspend fun close() = consumer.close()
 
     override suspend fun consume(key: UUID, value: String?) {
+        if (topic == Environment.DELTAKERLISTE_V2_TOPIC && !unleashToggle.skalLeseGjennomforingerV2()) {
+            return
+        }
+
         if (value == null) {
             deltakerlisteRepository.delete(key)
         } else {
